- Relocate Command IO from `org.grails.cli.commands.io` to `org.grails.cli.core.io` - Relocate Command Templates from `org.grails.cli.commands.templates` to `org.grails.cli.core.template`
