+++ grace-shell/src/main/groovy/org/grails/cli/commands/CreateAppCommand.groovy
@@ -55,7 +55,7 @@ import org.grails.build.logging.GrailsConsoleAntProject
 import org.grails.build.logging.GrailsConsoleLogger
 import org.grails.build.parsing.CommandLine
 import org.grails.cli.GrailsCli
-import org.grails.cli.commands.io.GradleDependency
+import org.grails.cli.core.io.GradleDependency
 import org.grails.cli.profile.Feature
 import org.grails.cli.profile.Profile
 import org.grails.cli.profile.ProfileRepository
