+++ grace-shell/src/main/groovy/org/grails/cli/core/io/FileSystemInteraction.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.commands.io
+package org.grails.cli.core.io
 
 import org.springframework.core.io.Resource
