+++ grace-shell/src/main/groovy/org/grails/cli/core/io/GradleDependency.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.commands.io
+package org.grails.cli.core.io
 
 import org.eclipse.aether.artifact.Artifact
 import org.eclipse.aether.graph.Dependency
