+++ grace-shell/src/main/groovy/org/grails/cli/core/template/TemplateRendererImpl.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.commands.templates
+package org.grails.cli.core.template
 
 import java.nio.charset.StandardCharsets
 
@@ -28,8 +28,8 @@ import org.springframework.core.io.ResourceLoader
 import grails.codegen.model.Model
 
 import org.grails.cli.commands.ExecutionContext
-import org.grails.cli.commands.io.FileSystemInteraction
-import org.grails.cli.commands.io.FileSystemInteractionImpl
+import org.grails.cli.core.io.FileSystemInteraction
+import org.grails.cli.core.io.FileSystemInteractionImpl
 import org.grails.cli.interactive.completers.ClassNameCompleter
 import org.grails.cli.profile.Profile
 import org.grails.cli.profile.ProfileRepository
