+++ grace-shell/src/main/groovy/org/grails/cli/profile/steps/RenderStep.groovy
@@ -24,7 +24,7 @@ import grails.util.GrailsNameUtils
 
 import org.grails.build.parsing.CommandLine
 import org.grails.cli.commands.ExecutionContext
-import org.grails.cli.commands.templates.SimpleTemplate
+import org.grails.cli.core.template.SimpleTemplate
 import org.grails.cli.interactive.completers.ClassNameCompleter
 import org.grails.cli.profile.AbstractStep
 import org.grails.cli.profile.Profile
