+++ grace-shell/src/main/groovy/org/grails/cli/profile/support/ArtefactVariableResolver.groovy
@@ -20,7 +20,7 @@ import groovy.transform.CompileStatic
 import grails.util.GrailsNameUtils
 
 import org.grails.cli.commands.ExecutionContext
-import org.grails.cli.commands.templates.SimpleTemplate
+import org.grails.cli.core.template.SimpleTemplate
 
 /**
  * @author Graeme Rocher
